add subpages and components
