+++ components/HomeSpotLight.js
@@ -1,0 +1,11 @@
+import { StyleHeadline } from "./ArtPieces";
+import Header from "./Header";
+
+export default function HomeSpotLight() {
+  return (
+    <>
+      <Header />
+      <h2>Picutre of the day</h2>
+    </>
+  );
+}
