+++ pages/Subpages/Spotlight.js
@@ -1,0 +1,12 @@
+import HomeSpotLight from "@/components/HomeSpotLight";
+
+export default function SpotLight() {
+
+    return (
+        <>
+        <HomeSpotLight/>
+        </>
+    )
+
+
+}
